update: fix issue when copying file

+++ src/main.rs
@@ -7,6 +7,7 @@ use std::{
     io::{self, ErrorKind, Stdout},
     path::{Path, PathBuf},
     process::Command,
+    time::{Duration, Instant},
 };
 use walkdir::WalkDir;
 
@@ -291,30 +292,33 @@ fn get_curr_path(path: String) -> String {
 }
 
 fn copy_dir_file_helper(src: &Path, new_src: &Path) -> anyhow::Result<()> {
-    let entries: Vec<_> = WalkDir::new(src)
-        .into_iter()
-        .filter_map(Result::ok)
-        .collect();
+    if src.is_file() {
+        fs::copy(src, new_src)?;
+    } else {
+        let entries: Vec<_> = WalkDir::new(src)
+            .into_iter()
+            .filter_map(Result::ok)
+            .collect();
+        entries.par_iter().try_for_each(|entry| {
+            let entry_path = entry.path();
+            let relative_path = entry_path.strip_prefix(src).unwrap();
+            let dst_path = new_src.join(relative_path);
 
-    entries.par_iter().try_for_each(|entry| {
-        let entry_path = entry.path();
-        let relative_path = entry_path.strip_prefix(src).unwrap();
-        let dst_path = new_src.join(relative_path);
-
-        if entry_path.is_dir() {
-            fs::create_dir_all(&dst_path)?;
-        } else if entry_path.is_file() {
-            if let Some(parent) = dst_path.parent() {
-                fs::create_dir_all(parent)?;
+            if entry_path.is_dir() {
+                fs::create_dir_all(&dst_path)?;
+            } else if entry_path.is_file() {
+                if let Some(parent) = dst_path.parent() {
+                    fs::create_dir_all(parent)?;
+                }
+                fs::copy(entry_path, dst_path)?;
+            } else {
+                println!("Error, file type not supported");
+                return Err(io::Error::new(ErrorKind::Other, "unsuported file type"));
             }
-            fs::copy(entry_path, dst_path)?;
-        } else {
-            println!("Error, file type not supported");
-            return Err(io::Error::new(ErrorKind::Other, "unsuported file type"));
-        }
 
-        Ok(())
-    })?;
+            Ok(())
+        })?;
+    }
 
     Ok(())
 }
@@ -375,6 +379,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                         Constraint::Length(3),
                         Constraint::Min(1),
                         Constraint::Length(3),
+                        Constraint::Length(1),
                     ]
                     .as_ref(),
                 )
@@ -416,12 +421,21 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                     InputMode::WatchRename => Style::default().fg(Color::Gray),
                 });
 
+
+            let mut list_title = String::new();
+            if app.loading {
+                let title_with_loader = format!("Copying Files...");
+                list_title.push_str(&title_with_loader);
+            } else {
+                list_title.push_str(&"List");
+            }
             // List of filtered items
             let list_block = List::new(filtered_items.clone())
                 .block(
                     Block::default()
                         .borders(Borders::ALL)
-                        .title("Filtered List"),
+                        .title(list_title.as_str()),
+                        //.title("Filtered List"),
                 )
                 .highlight_style(
                     Style::default()
@@ -467,17 +481,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             }
 
 
-            let loader_chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints([
-                    Constraint::Percentage(20),
-                    Constraint::Percentage(60),
-                    Constraint::Percentage(20),
-                ]. as_ref()
-                ).split(f.size());
-
-
-           // let loader_text = match
 
             f.render_widget(help_message, chunks[0]);
             f.render_widget(parsed_instructions.clone(), chunks[3]);
@@ -699,7 +702,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                         // item path to copy
                         if app.files.len() > 0 {
                             let index = state.selected();
-
+                            app.loading = true;
                             if let Some(indx) = index {
                                 // item to copy
                                 let selected_path = &app.files[indx];
@@ -716,10 +719,12 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                                     selected_path.to_string(),
                                     string_path.clone(),
                                 );
+
+                                app.input = src.to_str().unwrap().to_string();
                                 let new_src = Path::new(&new_path_with_new_name);
                                 copy_dir_file_helper(src, new_src)?;
                                 // show spinner that is downloading?
-
+                                app.loading = false;
                                 // TODO: create method that updates refreshes files
                                 match get_inner_files_info(string_path, app.show_hidden_files) {
                                     Ok(files) => {
